VMO-6541 fix type errors showing up during the build
